경고제거

+++ shikhondo/Enemy.h
@@ -27,7 +27,7 @@ protected:
 	FPOINT RandPos;
 	int RandNum;
 	int locationCount;
-	int missileSpeed;
+	float missileSpeed;
 	float checkTime;
 	float moveTime;
 	bool AutomaticMissile;
